fix: add clipboard fallback for insecure contexts (HTTP)
- Add clipboardWrite() utility with navigator.clipboard.writeText() + execCommand fallback
- Replace all navigator.clipboard.writeText() calls in app and ui packages
- Fixes copy button in HTTP/internal network environments
