+++ packages/ui/src/components/text-field.tsx
@@ -2,6 +2,7 @@ import { TextField as Kobalte } from "@kobalte/core/text-field"
 import { createSignal, Show, splitProps } from "solid-js"
 import type { ComponentProps } from "solid-js"
 import { useI18n } from "../context/i18n"
+import { clipboardWrite } from "../clipboard"
 import { IconButton } from "./icon-button"
 import { Tooltip } from "./tooltip"
 
@@ -69,7 +70,7 @@ export function TextField(props: TextFieldProps) {
 
   async function handleCopy() {
     const value = local.value ?? local.defaultValue ?? ""
-    await navigator.clipboard.writeText(value)
+    await clipboardWrite(value)
     setCopied(true)
     setTimeout(() => setCopied(false), 2000)
   }
